Vote feature with AJAX

+++ app/Repositories/VoteRepository.php
@@ -17,7 +17,10 @@ class VoteRepository {
 
     public function unvote(User $user, Ticket $ticket)
     {
+        if ( ! $user->hasVoted($ticket)) return false;
+
         $user->voted()->detach($ticket);
+        return true;
     }
 
 }
